requirements.txt
torch>=2.0.0
torchvision
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pillow>=10.0.0
trimesh>=4.0.0
numpy>=1.24.0
scipy>=1.10.0

# --- InstantMesh dependencies (install after cloning repo) ---
# diffusers>=0.24.0
# transformers>=4.35.0
# accelerate>=0.25.0
# safetensors>=0.4.0
# omegaconf
# einops
# rembg
# huggingface-hub

# Optional: better mesh processing
# pymeshlab
